upload form and data validation; column headers modal, home button on navbar; upload moved to its own file
