+++ server/README.md
@@ -95,20 +95,21 @@ note: recommended to use git bash inside vscode for all of this
 4. create a virtual environment in repo root (python -m venv venv)
 5. activate virtual environment (source venv/Scripts/activate or venv/bin/activate)
 6. install dependencies (`pip install -r requirements.txt -r requirements_dev.txt`)
-7. setting postgres:
+7. Start the db server: pg_ctl -D ~/pg/data/ -l logfile start
+8. setting postgres:
     - ```bash
         psql -U postgres -c "CREATE ROLE cpho_db_user with login"
         psql -U postgres -c "ALTER ROLE cpho_db_user createdb"
         createdb -U cpho_db_user cpho_dev_db
         ```
-8. seed the DB
+9. seed the DB
     - ```bash
         python ./manage.py migrate
         python ./manage.py loaddata cpho/fixtures/dimension_lookups.yaml
         python ./manage.py loaddata cpho/fixtures/periods.yaml
         python ./manage.py runscript cpho.scripts.dev
         ```
-9. `python manage.py runserver`
+10. `python manage.py runserver`
 
 ## Other useful commands:
