+++ server/cpho/views/__init__.py
@@ -1,2 +1,3 @@
 from .indicator_data import *
 from .indicators import *
+from .upload_indicators import *
