+++ server/cpho/views/indicators.py
@@ -1,5 +1,4 @@
 import csv
-import unicodedata
 from typing import Any
 
 from django import forms
@@ -60,17 +59,6 @@ class IndicatorForm(ModelForm):
     )
 
 
-class UploadForm(forms.Form):
-    csv_file = forms.FileField(
-        required=True,
-        widget=forms.FileInput(
-            attrs={
-                "class": "form-control",
-            }
-        ),
-    )
-
-
 class ListIndicators(ListView):
     model = Indicator
     template_name = "indicators/list_indicators.jinja2"
@@ -118,259 +106,3 @@ class EditIndicator(UpdateView):
         return {
             **super().get_context_data(**kwargs),
         }
-
-
-class UploadIndicator(FormView):
-    template_name = "indicators/upload_indicator.jinja2"
-    form_class = UploadForm
-    # success_url = "/upload_indicators/"
-
-    def get_success_url(self):
-        return reverse("list_indicators")
-
-    def form_valid(self, form):
-        print("VALID FORM")
-        print(self.get_success_url())
-        try:
-            csv_file = self.request.FILES["csv_file"]
-            if not csv_file.name.endswith(".csv"):
-                messages.error(self.request, tdt("File is not CSV type"))
-                raise forms.add_error(tdt("File is not CSV type"))
-                return HttpResponseRedirect(reverse("upload_indicator"))
-            if csv_file.multiple_chunks():
-                messages.error(
-                    self.request,
-                    tdt(
-                        "Uploaded file is too big. Maximum size allowed is 2 MB"
-                    ),
-                )
-                raise forms.ValidationError(
-                    tdt(
-                        "Uploaded file is too big. Maximum size allowed is 2 MB"
-                    )
-                )
-                return HttpResponseRedirect(reverse("upload_indicator"))
-
-            # ### Please dont remove this code; might need for custom parsing later
-
-            # file_data = csv_file.read().decode("utf-8-sig")
-            # lines = file_data.split("\n")
-            # data_dict = []
-            # for idx, line in enumerate(lines):
-            #     # print(line)
-            #     fields = line.split(",")
-            #     if idx == 0:
-            #         headers = [s.strip() for s in fields]
-            #         # TODO: check if headers are as expected
-            #     else:
-            #         if len(fields) != len(headers):
-            #             continue
-            #         data_row = {}
-            #         for i, field in enumerate(fields):
-            #             data_row[headers[i]] = field.strip()
-            #         data_dimension = self.deduce_dimension_type(data_row, idx)
-            #         print(data_dimension)
-            #         data_dict.append(data_row)
-            #         # print("SUCCESSFULLY ADDED ROW: ", idx)
-            # # print(data_dict)
-
-            file_data = csv_file.read().decode("utf-8-sig").splitlines()
-            # Not sure if this is needed in the future
-            # file_data = [
-            #     unicodedata.normalize("NFKD", line) for line in file_data
-            # ]
-            reader = csv.DictReader(file_data)
-            data_dict = []
-            for idx, data_row in enumerate(reader):
-                # Very annoying that i have to do this; but gives me errors otherwise
-                for key, value in data_row.items():
-                    data_row[key] = value.strip()
-                data_dimension = self.deduce_dimension_type(data_row, idx)
-                print(data_dimension)
-                data_dict.append(data_row)
-
-            messages.success(self.request, tdt("Data Uploaded Successfully"))
-
-        except Exception as err:
-            messages.error(
-                self.request, tdt("Unable to upload file: ") + repr(err)
-            )
-
-    def deduce_dimension_type(self, record, csv_line_number):
-        dimension_deduced_bool = False
-        dimension_deduced = None
-        dimension_col_name = None
-
-        checking_column = "Geography"
-        geography = record[checking_column]
-        if geography and geography.lower() not in ["canada", "null", ""]:
-            # print("DIMESNION: ", geography, " FOR: ", record, "\n")
-            dimension_deduced_bool = True
-            dimension_deduced = geography
-            dimension_col_name = checking_column
-
-        checking_column = "Sex"
-        sex = record[checking_column]
-        if sex and sex.lower() not in ["both sexes", "null", ""]:
-            # print("DIMESNION: ", sex, " FOR: ", record, "\n")
-            if dimension_deduced_bool:
-                print(
-                    "WARNING ON LINE (",
-                    csv_line_number,
-                    "): DIMENSION ALREADY DEDUCED ",
-                    dimension_deduced,
-                    "\nNEW DIMENSION FOUND: ",
-                    sex,
-                    "\nPLEASE CHECK RECORD: ",
-                    record,
-                )
-                return None
-            else:
-                dimension_deduced_bool = True
-                dimension_deduced = sex
-                dimension_col_name = checking_column
-
-        checking_column = "Gender"
-        gender = record[checking_column]
-        if gender and gender.lower() not in ["both genders", "null", ""]:
-            # print("DIMESNION: ", gender, " FOR: ", record, "\n")
-            if dimension_deduced_bool:
-                print(
-                    "WARNING ON LINE (",
-                    csv_line_number,
-                    "): DIMENSION ALREADY DEDUCED ",
-                    dimension_deduced,
-                    "\nNEW DIMENSION FOUND: ",
-                    gender,
-                    "\nPLEASE CHECK RECORD: ",
-                    record,
-                )
-                return None
-            else:
-                dimension_deduced_bool = True
-                dimension_deduced = gender
-                dimension_col_name = checking_column
-
-        checking_column = "Age_Group"
-        age_group = record[checking_column]
-        if age_group and age_group.lower() not in ["null", ""]:
-            # print("DIMESNION: ", age_group, " FOR: ", record, "\n")
-            if dimension_deduced_bool:
-                print(
-                    "WARNING ON LINE (",
-                    csv_line_number,
-                    "): DIMENSION ALREADY DEDUCED ",
-                    dimension_deduced,
-                    "\nNEW DIMENSION FOUND: ",
-                    age_group,
-                    "\nPLEASE CHECK RECORD: ",
-                    record,
-                )
-                return None
-            else:
-                dimension_deduced_bool = True
-                dimension_deduced = age_group
-                dimension_col_name = checking_column
-
-        checking_column = "Geography"
-        if not dimension_deduced_bool and geography.lower() == "canada":
-            # print("DIMESNION: ", geography, " FOR: ", record, "\n")
-            if dimension_deduced_bool:
-                print(
-                    "WARNING ON LINE (",
-                    csv_line_number,
-                    "): DIMENSION ALREADY DEDUCED ",
-                    dimension_deduced,
-                    "\nNEW DIMENSION FOUND: ",
-                    geography,
-                    "\nPLEASE CHECK RECORD: ",
-                    record,
-                )
-                return None
-            else:
-                dimension_deduced_bool = True
-                dimension_deduced = geography
-                dimension_col_name = checking_column
-
-        if not dimension_deduced_bool:
-            print(
-                "WARNING ON LINE (",
-                csv_line_number,
-                "): DIMENSION NOT FOUND ",
-                "\nPLEASE CHECK RECORD: ",
-                record,
-            )
-            return None
-        return {
-            "dimension_col_name": dimension_col_name,
-            "dimension_value": dimension_deduced,
-        }
-
-    def post(self, *args, **kwargs):
-        print("post")
-        # try:
-        #     csv_file = self.request.FILES["csv_file"]
-        #     if not csv_file.name.endswith(".csv"):
-        #         messages.error(self.request, tdt("File is not CSV type"))
-        #         return HttpResponseRedirect(reverse("upload_indicator"))
-        #     if csv_file.multiple_chunks():
-        #         messages.error(
-        #             self.request,
-        #             tdt(
-        #                 "Uploaded file is too big. Maximum size allowed is 2 MB"
-        #             ),
-        #         )
-        #         return HttpResponseRedirect(reverse("upload_indicator"))
-
-        #     # ### Please dont remove this code; might need for custom parsing later
-
-        #     # file_data = csv_file.read().decode("utf-8-sig")
-        #     # lines = file_data.split("\n")
-        #     # data_dict = []
-        #     # for idx, line in enumerate(lines):
-        #     #     # print(line)
-        #     #     fields = line.split(",")
-        #     #     if idx == 0:
-        #     #         headers = [s.strip() for s in fields]
-        #     #         # TODO: check if headers are as expected
-        #     #     else:
-        #     #         if len(fields) != len(headers):
-        #     #             continue
-        #     #         data_row = {}
-        #     #         for i, field in enumerate(fields):
-        #     #             data_row[headers[i]] = field.strip()
-        #     #         data_dimension = self.deduce_dimension_type(data_row, idx)
-        #     #         print(data_dimension)
-        #     #         data_dict.append(data_row)
-        #     #         # print("SUCCESSFULLY ADDED ROW: ", idx)
-        #     # # print(data_dict)
-
-        #     file_data = csv_file.read().decode("utf-8-sig").splitlines()
-        #     # Not sure if this is needed in the future
-        #     # file_data = [
-        #     #     unicodedata.normalize("NFKD", line) for line in file_data
-        #     # ]
-        #     reader = csv.DictReader(file_data)
-        #     data_dict = []
-        #     for idx, data_row in enumerate(reader):
-        #         # Very annoying that i have to do this; but gives me errors otherwise
-        #         for key, value in data_row.items():
-        #             data_row[key] = value.strip()
-        #         data_dimension = self.deduce_dimension_type(data_row, idx)
-        #         # print(data_dimension)
-        #         data_dict.append(data_row)
-
-        #     messages.success(self.request, tdt("Data Uploaded Successfully"))
-
-        # except Exception as err:
-        #     messages.error(
-        #         self.request, tdt("Unable to upload file: ") + repr(err)
-        #     )
-        super().post(*args, **kwargs)
-        form = self.get_form()
-        if form.is_valid():
-            return HttpResponseRedirect(self.get_success_url())
-        else:
-            print("form not valid")
-            print(form.errors)
-            return self.form_invalid(form)
