+++ server/static/cpho.css
@@ -12,3 +12,19 @@ footer {
 #footer-placeholder {
     height: 3rem;
 }
+
+.navbar h1 {
+    color: white;
+    margin: 0;
+    margin-right: 5px;
+}
+
+.navbar a.dropdown-item,
+.navbar a.dropdown-toggle {
+    text-decoration: none;
+}
+
+h1 abbr {
+    text-decoration: none !important;
+    border-bottom: none;
+}
